requirements.txt
Django==1.9.2
MySQL-python==1.2.5
PyYAML==3.12
django-extensions==1.6.1
django-rest-swagger==0.3.5
requests==2.11.1
djangorestframework==3.3.2
pyotp==2.2.1
six==1.10.0
django-user-roles
awscli
fuzzywuzzy
mandrill
rsa==3.4.2
XlsxWriter==0.9.3
jsonfield
django-celery
boto3==1.4.1
celery[redis]==4.2.1
python-gcm==0.4
pandas==0.19.1
PyMySQL==0.7.9
apns==2.0.1
djangorestframework-csv==2.0.0
suds==0.4
xhtml2pdf==0.0.6
wkhtmltopdf==0.2
reportlab==3.3.0
html5lib==1.0b8
num2words==0.5.4
lxml==3.7.3
gevent==1.2.1
zc.lockfile==1.2.1
facebook-sdk==2.0.0
google-api-python-client==1.6.2
httplib2==0.10.3
python-linkedin==4.1
oauth2client==3.0.0
oauthlib==2.0.2
xmltodict==0.11.0
xlwt
pyopenssl
xlrd==1.1.0
bleach==1.5.0
pycrypto==2.6.1
analytics-python
django-redis==3.8.4
django-redis-cache==1.7.1
elastic-apm==5.1.2
python-magic==0.4.15
